requirements.txt
pynacl
transformers
requests
